Fixed terrible MemoryArena bug:
The MemoryArena had a bug for almost 1 year where a new page was
allocated for every memory allocation after the first few. This resulted
in a massive increase in memory usage, and for things that got
allocated/de-allocated every frame, it basically meant a memory leak.

Just a small adjustment and everything seems now good.

Added also a variable to keep track of how many bytes are currently
allocated for debug/diagnostic reasons.