C-linkage removed. Some functions refactored:
I decided to remove C-linkage, because many functions in lsString are a
lot easier to use when function overloading is available.

Because of this, all the *BreakBy* functions have been removed and
replaced by a bunch of ls_unistrSplit() overloaded functions that allow
to split on a certain set of characters.

Also a new function has been added:
ls_unistrSeparateByNumber() which splits BUT keeps the delimeter (any
number in this case) as a string in the returned array.