Removed large overhead in Memory System.
Because most large allocations (> 1 page) are for items that are not
usually freed, like assets, there's no point in allowing them to be
sliced. The original implementation would base the max slices of a block
relative to the allocation size, which would mean these large blocks
would have hundred of thousand of slices, which at 24 bytes per slice
would start occupying a considerable amount of memory, completely
uselessly, since the blocks would never be sliced.